Add GitHub CI setup
As Travis CI is now severily limited for OSS projects and I would say GitHub Actions are more accessible for GitHub projects.

Based on a modified copy of https://raw.githubusercontent.com/sindresorhus/eslint-plugin-unicorn/master/.github/workflows/ci.yml

Adds tests for Node 14 and 15 on top of the 10 and 12 that's in the Travis file.