Add `GlobalThis` type
This type is for times when you need to access non-standard properties off of `window`, `self`, or `globalThis`. For example:

```ts
import type {GlobalThis} from 'type-fest';

interface ExtraProps extends GlobalThis {
  readonly GLOBAL_TOKEN: string;
}

const extraProps = window as ExtraProps;
```

---

Normally, you would define a declaration file to add global variables.

When you need to do this in a more targeted manner though, a common approach is:

```ts
interface ExtraProps extends Window {
  readonly GLOBAL_TOKEN: string
}

const extraProps = window as unknown as ExtraProps;
```

Without the cast through `unknown`, you get error `TS2352`.

You can try to workaround that with the following:

```ts
type GlobalThis = (typeof globalThis) & Window;
interface ExtraProps extends GlobalThis {
  readonly GLOBAL_TOKEN: string
}
```

But then you get the error: `Property 'Infinity' of type 'number' is not assignable to numeric index type 'Window'.(2412)` (also for `NaN`).

So you have to redefine `Infinity` and `NaN` as `never` to get this working.

---

Note: In environments that choose to not load the `"dom"` lib (e.g., for workers or node), the global `Window` interface will not be defined. The only way to modify `Infinity` and `NaN` would be to declare `Window` in the `global` scope. This is undesirable though as now an empty `Window` symbol will be present for users of `type-fest`. The workaround here is to make use of a `@ts-ignore` to treat `Window` as `any` if it is not present.