Add `IsUnion` type
Permits detecting if a given type is a singular type or a union of types. Useful when you only want to accept a union of types or a single type.

This commit uses a pattern where, instead of returning `true` and `false`, the caller can specify the result in the true and false cases. Effectively, you can write:

```ts
IsUnion<T, IfTrue, IfFalse>
```

instead of:

```ts
IsUnion<T> extends true ? IfTrue : IfFalse
```