Add `Destructurable` type
## Overview
This pull request introduces the `Destructurable` type that aims to resolve an issue encountered when trying to destructure the return values of functions with union types.

## TypeScript Compiler's Type Inference
When we define a function and omit the return type, the TypeScript compiler infers the return type automatically. For example, in the provided code snippet:

```typescript
function doSomething() {
  return Math.random() > 0.5
    ? { type: "success" as const, value: "foo" }
    : { type: "failure" as const, error: new Error("Something wrong") };
}
```

The compiler infers the `ReturnType` as:

```typescript
type ReturnType =
  | {
      type: "success";
      value: string;
      error?: undefined;
    }
  | {
      type: "failure";
      error: Error;
      value?: undefined;
    };
```

Thanks to this inferred type, we can destructure the return value directly as shown:

```typescript
const { type, value, error } = doSomething();
```

## Problem

However, when defining functions with union types as return values, the TypeScript compiler is unable to destructure the return values directly, causing errors when accessing properties of the return value. For instance, when explicitly annotating the return type of the `doSomething()` function with `Success | Failure` union type:

```typescript
type Success = {
  type: "success";
  value: string;
};
type Failure = {
  type: "failure";
  error: Error;
};
function doSomething(): Success | Failure {
  return Math.random() > 0.5
    ? { type: "success" as const, value: "foo" }
    : { type: "failure" as const, error: new Error("Something wrong") };
}
const { type, value, error } = doSomething();
```

It leads to the following errors:

> Property 'value' does not exist on type 'Success | Failure'.(2339)
> Property 'error' does not exist on type 'Success | Failure'.(2339)

A workaround for this issue would be to add the error property to the Success type and the value property to the Failure type. However, this approach adds unnecessary properties to the types.

## Solution

The `Destructurable` type resolves the problem by allowing the union types to be destructured directly. In the example provided, the `Result` type is defined as a `Destructurable<Success | Failure>`:

```typescript
type Result = Destructurable<Success | Failure>;
function doSomething(): Result {
  return Math.random() > 0.5
    ? { type: "success", value: "foo" }
    : { type: "failure", error: new Error("Something wrong") };
}
```

The `Destructurable<Success | Failure>` type is calculated as:

```typescript
type ReturnType =
  | {
      type: "success";
      value: string;
      error?: undefined;
    }
  | {
      type: "failure";
      error: Error;
      value?: undefined;
    };
```

With the `Destructurable` type, the result can be destructured directly, and narrowing also works as expected:

```typescript
const { type, value, error } = doSomething();

if (type === "success") {
  value.toUpperCase();
} else {
  error.message;
}
```

This new type simplifies the handling of union types in functions' return values, making it easier to destructure and work with the properties of those types.