Create grouped dependabot.yml
Fixes #840

I know @sindresorhus thinks this can get noisy, but I tried to make it as non-noisy as possible.

I would really love to get notified when eg. a new `tsd`, `xo` or similar dependency needs to be updated, and same for GitHub Actions.

This setup makes it so that at max one PR for npm and one PR for GitHub Actions will be created, and it will only be created/updated weekly, and for npm it will only create it if the version range _needs_ to be increased, so no needless bumping of the lower range of a version.